Initial Commit for the Base ErrorChecker Class
This class is not in any way completed, and is only a barebone for the entire branch.

The @Deprecated items in the class are meant for workflow understanding, and in no way are recomended for use ANYWHERE.

Note: All of this is related to everything else being pushed around this time. If something does not make sense, read everything first.